Atomic database upgrades (#1892)

+++ src/services/repositories/metadata.ts
@@ -13,7 +13,7 @@ class MetadataRepo {
 
   initialized: boolean;
 
-  async init() {
+  async init(cleanDbs = true) {
     const conn = await new Connection(dbMan.baseConnConfig).connect();
     try {
       await conn.query(`CREATE DATABASE ${this.database};`);
@@ -38,15 +38,15 @@ class MetadataRepo {
     this.dbRepo = this.conn.getRepository(IasqlDatabase);
     // In case of partially-failed disconnects, we delete all IasqlDatabase records that don't
     // actually have a database on startup.
-    const expectedDbs = await this.dbRepo.find();
-    const actualDbs = (
-      await this.conn.query(`
-      SELECT datname FROM pg_database;
-    `)
-    ).map((r: any) => r.datname);
-    for (const expectedDb of expectedDbs) {
-      if (actualDbs.includes(expectedDb.pgName)) continue;
-      await this.dbRepo.remove(expectedDb);
+    if (cleanDbs) {
+      const expectedDbs = await this.dbRepo.find();
+      const actualDbs = (await this.conn.query('SELECT datname FROM pg_database;')).map(
+        (r: any) => r.datname,
+      );
+      for (const expectedDb of expectedDbs) {
+        if (actualDbs.includes(expectedDb.pgName)) continue;
+        await this.dbRepo.remove(expectedDb);
+      }
     }
     this.initialized = true;
   }
